ref(chat): Refactor Chat components
Remove Abstract component
Convert web component to function component
