+++ react/features/chat/components/web/Chat.tsx
@@ -1,16 +1,15 @@
 import clsx from 'clsx';
-import React from 'react';
+import React, { useCallback } from 'react';
 import { connect } from 'react-redux';
 
+import { IReduxState } from '../../../app/types';
 import { translate } from '../../../base/i18n/functions';
+import { getLocalParticipant } from '../../../base/participants/functions';
 import Tabs from '../../../base/ui/components/web/Tabs';
 import PollsPane from '../../../polls/components/web/PollsPane';
-import { toggleChat } from '../../actions.web';
-import { CHAT_TABS } from '../../constants';
-import AbstractChat, {
-    IProps,
-    _mapStateToProps
-} from '../AbstractChat';
+import { sendMessage, setIsPollsTabFocused, toggleChat } from '../../actions.web';
+import { CHAT_TABS, SMALL_WIDTH_THRESHOLD } from '../../constants';
+import { IProps as AbstractProps } from '../../types';
 
 import ChatHeader from './ChatHeader';
 import ChatInput from './ChatInput';
@@ -19,75 +18,101 @@ import KeyboardAvoider from './KeyboardAvoider';
 import MessageContainer from './MessageContainer';
 import MessageRecipient from './MessageRecipient';
 
-/**
- * React Component for holding the chat feature in a side panel that slides in
- * and out of view.
- */
-class Chat extends AbstractChat<IProps> {
+interface IProps extends AbstractProps {
 
     /**
-     * Reference to the React Component for displaying chat messages. Used for
-     * scrolling to the end of the chat messages.
+     * Whether the chat is opened in a modal or not (computed based on window width).
      */
-    _messageContainerRef: Object;
+    _isModal: boolean;
 
     /**
-     * Initializes a new {@code Chat} instance.
+     * True if the chat window should be rendered.
+     */
+    _isOpen: boolean;
+
+    /**
+     * True if the polls feature is enabled.
+     */
+    _isPollsEnabled: boolean;
+
+    /**
+     * Whether the poll tab is focused or not.
+     */
+    _isPollsTabFocused: boolean;
+
+    /**
+     * Number of unread poll messages.
+     */
+    _nbUnreadPolls: number;
+
+    /**
+     * Function to send a text message.
      *
-     * @param {Object} props - The read-only properties with which the new
-     * instance is to be initialized.
+     * @protected
      */
-    constructor(props: IProps) {
-        super(props);
-
-        this._messageContainerRef = React.createRef();
-
-        // Bind event handlers so they are only bound once for every instance.
-        this._onChatTabKeyDown = this._onChatTabKeyDown.bind(this);
-        this._onEscClick = this._onEscClick.bind(this);
-        this._onPollsTabKeyDown = this._onPollsTabKeyDown.bind(this);
-        this._onToggleChat = this._onToggleChat.bind(this);
-        this._onChangeTab = this._onChangeTab.bind(this);
-    }
+    _onSendMessage: Function;
 
     /**
-     * Implements React's {@link Component#render()}.
-     *
-     * @inheritdoc
-     * @returns {ReactElement}
+     * Function to toggle the chat window.
      */
-    render() {
-        const { _isOpen, _isPollsEnabled, _showNamePrompt } = this.props;
-
-        return (
-            _isOpen ? <div
-                className = 'sideToolbarContainer'
-                id = 'sideToolbarContainer'
-                onKeyDown = { this._onEscClick } >
-                <ChatHeader
-                    className = 'chat-header'
-                    isPollsEnabled = { _isPollsEnabled }
-                    onCancel = { this._onToggleChat } />
-                { _showNamePrompt
-                    ? <DisplayNameForm isPollsEnabled = { _isPollsEnabled } />
-                    : this._renderChat() }
-            </div> : null
-        );
-    }
+    _onToggleChat: Function;
 
     /**
-     * Key press handler for the chat tab.
+     * Function to display the chat tab.
      *
-     * @param {KeyboardEvent} event - The event.
-     * @returns {void}
+     * @protected
      */
-    _onChatTabKeyDown(event: React.KeyboardEvent) {
-        if (event.key === 'Enter' || event.key === ' ') {
-            event.preventDefault();
-            event.stopPropagation();
-            this._onToggleChatTab();
-        }
-    }
+    _onToggleChatTab: Function;
+
+    /**
+     * Function to display the polls tab.
+     *
+     * @protected
+     */
+    _onTogglePollsTab: Function;
+
+    /**
+     * Whether or not to block chat access with a nickname input form.
+     */
+    _showNamePrompt: boolean;
+}
+
+const Chat = ({
+    _isModal,
+    _isOpen,
+    _isPollsEnabled,
+    _isPollsTabFocused,
+    _messages,
+    _nbUnreadMessages,
+    _nbUnreadPolls,
+    _onSendMessage,
+    _onToggleChat,
+    _onToggleChatTab,
+    _onTogglePollsTab,
+    _showNamePrompt,
+    dispatch,
+    t
+}: IProps) => {
+    /**
+    * Sends a text message.
+    *
+    * @private
+    * @param {string} text - The text message to be sent.
+    * @returns {void}
+    * @type {Function}
+    */
+    const onSendMessage = useCallback((text: string) => {
+        dispatch(sendMessage(text));
+    }, []);
+
+    /**
+    * Toggles the chat window.
+    *
+    * @returns {Function}
+    */
+    const onToggleChat = useCallback(() => {
+        dispatch(toggleChat());
+    }, []);
 
     /**
      * Click handler for the chat sidenav.
@@ -95,27 +120,23 @@ class Chat extends AbstractChat<IProps> {
      * @param {KeyboardEvent} event - Esc key click to close the popup.
      * @returns {void}
      */
-    _onEscClick(event: React.KeyboardEvent) {
-        if (event.key === 'Escape' && this.props._isOpen) {
+    const _onEscClick = useCallback((event: React.KeyboardEvent) => {
+        if (event.key === 'Escape' && _isOpen) {
             event.preventDefault();
             event.stopPropagation();
-            this._onToggleChat();
+            onToggleChat();
         }
-    }
+    }, []);
 
     /**
-     * Key press handler for the polls tab.
+     * Change selected tab.
      *
-     * @param {KeyboardEvent} event - The event.
+     * @param {string} id - Id of the clicked tab.
      * @returns {void}
      */
-    _onPollsTabKeyDown(event: React.KeyboardEvent) {
-        if (event.key === 'Enter' || event.key === ' ') {
-            event.preventDefault();
-            event.stopPropagation();
-            this._onTogglePollsTab();
-        }
-    }
+    const _onChangeTab = useCallback((id: string) => {
+        dispatch(setIsPollsTabFocused(id !== CHAT_TABS.CHAT));
+    }, []);
 
     /**
      * Returns a React Element for showing chat messages and a form to send new
@@ -124,12 +145,10 @@ class Chat extends AbstractChat<IProps> {
      * @private
      * @returns {ReactElement}
      */
-    _renderChat() {
-        const { _isPollsEnabled, _isPollsTabFocused } = this.props;
-
+    function _renderChat() {
         return (
             <>
-                { _isPollsEnabled && this._renderTabs() }
+                {_isPollsEnabled && _renderTabs()}
                 <div
                     aria-labelledby = { CHAT_TABS.CHAT }
                     className = { clsx(
@@ -141,12 +160,12 @@ class Chat extends AbstractChat<IProps> {
                     role = 'tabpanel'
                     tabIndex = { 0 }>
                     <MessageContainer
-                        messages = { this.props._messages } />
+                        messages = { _messages } />
                     <MessageRecipient />
                     <ChatInput
-                        onSend = { this._onSendMessage } />
+                        onSend = { onSendMessage } />
                 </div>
-                { _isPollsEnabled && (
+                {_isPollsEnabled && (
                     <>
                         <div
                             aria-labelledby = { CHAT_TABS.POLLS }
@@ -169,13 +188,11 @@ class Chat extends AbstractChat<IProps> {
      * @private
      * @returns {ReactElement}
      */
-    _renderTabs() {
-        const { _isPollsEnabled, _isPollsTabFocused, _nbUnreadMessages, _nbUnreadPolls, t } = this.props;
-
+    function _renderTabs() {
         return (
             <Tabs
                 accessibilityLabel = { t(_isPollsEnabled ? 'chat.titleWithPolls' : 'chat.title') }
-                onChange = { this._onChangeTab }
+                onChange = { _onChangeTab }
                 selected = { _isPollsTabFocused ? CHAT_TABS.POLLS : CHAT_TABS.CHAT }
                 tabs = { [ {
                     accessibilityLabel: t('chat.tabs.chat'),
@@ -194,24 +211,51 @@ class Chat extends AbstractChat<IProps> {
         );
     }
 
-    /**
-    * Toggles the chat window.
-    *
-    * @returns {Function}
-    */
-    _onToggleChat() {
-        this.props.dispatch(toggleChat());
-    }
+    return (
+        _isOpen ? <div
+            className = 'sideToolbarContainer'
+            id = 'sideToolbarContainer'
+            onKeyDown = { _onEscClick } >
+            <ChatHeader
+                className = 'chat-header'
+                isPollsEnabled = { _isPollsEnabled }
+                onCancel = { onToggleChat } />
+            {_showNamePrompt
+                ? <DisplayNameForm isPollsEnabled = { _isPollsEnabled } />
+                : _renderChat()}
+        </div> : null
+    );
+};
 
-    /**
-     * Change selected tab.
-     *
-     * @param {string} id - Id of the clicked tab.
-     * @returns {void}
-     */
-    _onChangeTab(id: string) {
-        id === CHAT_TABS.CHAT ? this._onToggleChatTab() : this._onTogglePollsTab();
-    }
+/**
+ * Maps (parts of) the redux state to {@link Chat} React {@code Component}
+ * props.
+ *
+ * @param {Object} state - The redux store/state.
+ * @param {any} _ownProps - Components' own props.
+ * @private
+ * @returns {{
+ *     _isOpen: boolean,
+ *     _messages: Array<Object>,
+ *     _showNamePrompt: boolean
+ * }}
+ */
+function _mapStateToProps(state: IReduxState, _ownProps: any) {
+    const { isOpen, isPollsTabFocused, messages, nbUnreadMessages } = state['features/chat'];
+    const { nbUnreadPolls } = state['features/polls'];
+    const _localParticipant = getLocalParticipant(state);
+    const { disablePolls } = state['features/base/config'];
+
+    return {
+        _isModal: window.innerWidth <= SMALL_WIDTH_THRESHOLD,
+        _isOpen: isOpen,
+        _isPollsEnabled: !disablePolls,
+        _isPollsTabFocused: isPollsTabFocused,
+        _messages: messages,
+        _nbUnreadMessages: nbUnreadMessages,
+        _nbUnreadPolls: nbUnreadPolls,
+        _showNamePrompt: !_localParticipant?.name
+    };
 }
 
 export default translate(connect(_mapStateToProps)(Chat));
